requirements.txt
Flask==2.1.1
pandas==1.3.5
nixtla==0.6.0
google-cloud-storage==2.9.0